Added proper message is overdue checks. Abstracted duration and state to instance abstract
