+++ src/instances/MetagameTerritoryInstance.ts
@@ -6,7 +6,6 @@ import PS2AlertsInstanceInterface from '../interfaces/PS2AlertsInstanceInterface
 import {Ps2alertsEventState} from '../ps2alerts-constants/ps2alertsEventState';
 import {Bracket} from '../ps2alerts-constants/bracket';
 import TerritoryResultInterface from '../interfaces/TerritoryResultInterface';
-import moment from 'moment/moment';
 
 export default class MetagameTerritoryInstance extends InstanceAbstract implements PS2AlertsInstanceInterface {
     public bracket?: Bracket;
@@ -19,8 +18,8 @@ export default class MetagameTerritoryInstance extends InstanceAbstract implemen
         public readonly zone: Zone,
         public readonly censusInstanceId: number,
         public readonly censusMetagameEventType: MetagameEventType,
-        public readonly duration: number, // Stored in Milliseconds
-        public state: Ps2alertsEventState,
+        duration: number, // Stored in Milliseconds
+        state: Ps2alertsEventState,
         bracket?: Bracket,
     ) {
         super(
@@ -29,6 +28,8 @@ export default class MetagameTerritoryInstance extends InstanceAbstract implemen
             timeStarted,
             timeEnded,
             result,
+            duration,
+            state,
         );
         this.bracket = bracket ?? undefined;
     }
@@ -38,18 +39,5 @@ export default class MetagameTerritoryInstance extends InstanceAbstract implemen
             ? world === this.world && zone === this.zone
             : world ? world === this.world
             : zone === this.zone;
-    }
-
-    public overdue(): boolean {
-        // If now in milliseconds is greater than start time + duration
-        return Date.now() > (this.timeStarted.getTime() + this.duration);
-    }
-
-    // Returns the current second tick of the alert
-    public currentDuration(): number {
-        // Return current difference in seconds between start and now
-        const nowUnix = moment().unix() * 1000;
-        // Holy mother of brackets batman!
-        return parseInt(((nowUnix - this.timeStarted.getTime()) / 1000).toFixed(0), 10);
     }
 }
