+++ src/interfaces/PS2AlertsInstanceInterface.ts
@@ -18,6 +18,8 @@ export default interface PS2AlertsInstanceInterface {
 
     overdue(): boolean;
 
+    messageOverdue(timestamp: Date): boolean;
+
     match(world: World | null, zone: Zone | null): boolean;
 
     currentDuration(): number;
